tidy: Rename optional/ -> ui/, and split statusbars.lua
