--- config/nvim/lua/ide/init.lua
@@ -1,9 +1,0 @@
-require('ide.file_explorer')
-require('ide.lsp')
-require('ide.completion')
-require('ide.diagnostics')
-require('ide.git')
-require('ide.linters')
-require('ide.statusbars')
-require('ide.telescope')
-require('ide.treesitter')
