--- config/nvim/lua/optional/init.lua
@@ -1,4 +1,0 @@
-require('optional.colorizer')
-require('optional.glow')
-require('optional.nvim_navic')
-require('optional.todo_comments')
