+++ config/nvim/lua/ui/nvim_navic.lua
